split detection util in lib/api/skill_detection
